# weft/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Weft API
  version: 0.2.0
  description: |
    Coordinator service for Loominal multi-agent infrastructure.

    Weft routes work, manages agent lifecycle, handles scaling, and provides a REST API
    for fleet management. It runs as a multi-tenant service that handles multiple projects
    in a single deployment.

    ## Key Features
    - **Agent Management**: Register, discover, and shutdown agents
    - **Work Distribution**: Submit and track work items across agents
    - **Spin-Up Targets**: Manage dynamic agent provisioning
    - **Batch Operations**: Perform bulk operations on agents, targets, and work
    - **Pagination**: All list endpoints support cursor-based pagination
    - **Semantic Context**: Responses include resolved agent details for better decision-making

  contact:
    name: Michael LoPresti
    email: mike@lopresti.org

  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://weft.loominal.dev
    description: Production server
...
paths:
...
  /api/agents:
    get:
      summary: List registered agents
      description: |
        Returns a paginated list of registered agents with optional filtering.
        Supports cursor-based pagination for efficient traversal of large agent fleets.
      operationId: listAgents
      tags:
        - Agents
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          description: Filter by agent type
          schema:
            $ref: '#/components/schemas/AgentType'
        - name: status
          in: query
          description: Filter by agent status
          schema:
            $ref: '#/components/schemas/AgentStatus'
        - name: capability
          in: query
          description: Filter by capability
          schema:
            type: string
            example: typescript
        - name: cursor
          in: query
          description: Pagination cursor from previous response
          schema:
            type: string
            example: eyJvZmZzZXQiOjUwLCJsaW1pdCI6NTB9
        - name: limit
          in: query
          description: Maximum number of agents to return per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: List of agents with pagination metadata
          content:
            application/json:
              schema:
                type: object
                required:
                  - agents
                  - count
                  - hasMore
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/RegisteredAgent'
                  count:
                    type: integer
                    description: Number of agents in this page
                    example: 50
                  total:
                    type: integer
                    description: Total number of agents matching filter
                    example: 847
                  hasMore:
                    type: boolean
                    description: Whether more results are available
                    example: true
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor for fetching the next page
                    example: eyJvZmZzZXQiOjUwLCJsaW1pdCI6NTB9
                  prevCursor:
                    type: string
                    nullable: true
                    description: Cursor for fetching the previous page
                    example: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/agents/{guid}:
    get:
      summary: Get agent details
      description: Returns detailed information about a specific agent by GUID
      operationId: getAgent
      tags:
        - Agents
      security:
        - bearerAuth: []
      parameters:
        - name: guid
          in: path
          required: true
          description: Agent GUID
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredAgent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/agents/{guid}/shutdown:
    post:
      summary: Request agent shutdown
      description: |
        Requests a specific agent to shutdown. Supports graceful shutdown
        where the agent completes current work before terminating.
      operationId: shutdownAgent
      tags:
        - Agents
      security:
        - bearerAuth: []
      parameters:
        - name: guid
          in: path
          required: true
          description: Agent GUID to shutdown
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                graceful:
                  type: boolean
                  default: true
                  description: Wait for current work to complete before shutdown
                  example: true
      responses:
        '200':
          description: Shutdown request sent successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                  - graceful
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Shutdown request sent to agent 550e8400-e29b-41d4-a716-446655440000
                  graceful:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/agents/shutdown-batch:
    post:
      summary: Batch shutdown agents
      description: |
        Shuts down multiple agents matching filter criteria or specific GUIDs.
        Useful for scaling down agent fleets or maintenance operations.
      operationId: batchShutdownAgents
      tags:
        - Agents
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchShutdownRequest'
            examples:
              filterBased:
                summary: Shutdown idle agents
                value:
                  filter:
                    status: idle
                    idleTimeMs: 300000
                  graceful: true
                  reason: scaling-down
              guidBased:
                summary: Shutdown specific agents
                value:
                  agentGuids:
                    - 550e8400-e29b-41d4-a716-446655440000
                    - 6ba7b810-9dad-11d1-80b4-00c04fd430c8
                  graceful: true
      responses:
        '200':
          description: Batch operation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchShutdownResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication (configure via API_TOKENS env var)
  schemas:
    AgentType:
      type: string
      enum:
        - copilot-cli
        - claude-code
      description: Type of AI agent
    AgentStatus:
      type: string
      enum:
        - online
        - busy
        - offline
      description: Current status of an agent
...
    RegisteredAgent:
      type: object
      required:
        - guid
        - handle
        - agentType
        - status
        - capabilities
        - boundaries
        - hostname
        - projectId
        - visibility
        - currentTaskCount
        - maxConcurrentTasks
        - spindownAfterIdleMs
        - lastHeartbeat
        - lastActivity
        - registeredAt
      properties:
        guid:
          type: string
          format: uuid
        handle:
          type: string
        agentType:
          $ref: '#/components/schemas/AgentType'
        status:
          $ref: '#/components/schemas/AgentStatus'
        capabilities:
          type: array
          items:
            type: string
        boundaries:
          type: array
          items:
            type: string
          description: Workload boundaries this agent accepts
        hostname:
          type: string
        projectId:
          type: string
        username:
          type: string
        visibility:
          type: string
          enum: [private, project-only, user-only, public]
        currentTaskCount:
          type: integer
        maxConcurrentTasks:
          type: integer
        spindownAfterIdleMs:
          type: integer
        lastHeartbeat:
          type: string
          format: date-time
        lastActivity:
          type: string
          format: date-time
        registeredAt:
          type: string
          format: date-time
...
    BatchShutdownRequest:
      type: object
      properties:
        filter:
          type: object
          description: Filter criteria for selecting agents
          properties:
            status:
              $ref: '#/components/schemas/AgentStatus'
            idleTimeMs:
              type: integer
            agentType:
              $ref: '#/components/schemas/AgentType'
            boundary:
              type: string
            capability:
              type: string
        agentGuids:
          type: array
          items:
            type: string
            format: uuid
          description: Specific agent GUIDs to shutdown
        graceful:
          type: boolean
          default: true
        gracePeriodMs:
          type: integer
          default: 30000
        reason:
          type: string
          enum: [manual, scaling-down, maintenance, error]
    BatchShutdownResponse:
      type: object
      required:
        - success
        - failed
        - count
        - errors
        - completedAt
        - totalProcessed
        - successRate
        - shutdownAgents
        - graceful
      properties:
        success:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            type: string
        count:
          type: integer
        errors:
          type: object
          additionalProperties:
            type: string
        completedAt:
          type: string
          format: date-time
        totalProcessed:
          type: integer
        successRate:
          type: number
          format: float
        shutdownAgents:
          type: array
          items:
            type: string
        graceful:
          type: boolean
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: ValidationError
            message: Invalid request parameters
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or missing authentication token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: InternalServerError
            message: An unexpected error occurred